Refactor Audio Feature Casting Path

+++ src/transformers/models/musicflamingo/modeling_musicflamingo.py
@@ -606,8 +606,7 @@ class MusicFlamingoForConditionalGeneration(MusicFlamingoPreTrainedModel, Genera
             `torch.FloatTensor`:
                 The audio embeddings.
         """
-        # Encode audio with dtype conversion and audio_times
-        input_features = input_features.to(dtype=self.audio_tower.conv1.weight.dtype)
+        # Encode audio with audio_times
         encoder_output = self.audio_tower(
             input_features, input_features_mask=input_features_mask, audio_times=audio_times
         )
